Mark delegate methods as optional
Marking a protocol and its methods as @objc makes the default conformance extension to be ignored. Marking the default conformance extension as @objc does not work. Making the methods optional instead (which is now available, as the methods are marked as @objc) fixes this, providing the same results as before when using a delegate without implementing all methods.